Split the test and script families their own files already named
Three maintainability findings, each fixed the way the repository's own
notes said to fix it.

presentation.test.ts sat at 496 lines against a 500-line transition
exception whose reason read "row and column building are the next families
out". Those two families are now presentation.rows.test.ts and
presentation.columns.test.ts, the original is 262 lines, and the exception
is deleted rather than raised. 152 tests before, 152 after.

docs-header.ts had two nearly identical sibling walks four levels of nesting
apart -- one scans forward past other elements until it finds a table, the
other tags only the immediately following element. Close enough to look the
same, far enough apart that the difference was invisible. Both are now named
functions, and the per-cell column labelling is one too. That pushed the file
over its own 360-line cap, whose reason read "split search vs toc vs tables
next", so the tables third moves to docs-tables.ts and the cap drops to 280.

ToolForm.tsx had an options `.map` holding a button holding a click handler,
eight braces deep inside the field `.map` inside the component. The listbox
rows move to FieldSelectOptions; the state stays put.

ResultTable rendered two near-identical `<td>` branches that differed only in
class and content, each carrying its own copy of a seven-argument
context-menu handler. Now one.

Two nesting sites are left and are deliberate. ResultTable's rows-by-columns
double `.map` is what a table is; extracting a cell component would move the
depth to another file behind seven props. os-tabs.ts's depth is a `.then`
inside a feature-detect branch inside a handler -- the shape of an optional
async upgrade, and already commented as such.

tsc, eslint, 152 unit tests, astro check (0 errors, 54 files), site build,
a11y (0 violations, both themes), and the file-size guard all pass.
